The prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -1,15 +1,15 @@
 #include "main.h"
 #include <stdio.h>
+#include <math.h>
 /**
  * main - finds and prints the largest prime factor of the number
- *
  *
  * Return: 0 if success
  */
 
 int main(void)
 {
-	long long int num = 612852475143, i;
+	long int num = 612852475143, i;
 
 	for (i = 2; i <= num; i++)
 	{
@@ -19,6 +19,6 @@ int main(void)
 			i = i - 1;
 		}
 	}
-	printf("%lld\n", i);
+	printf("%ld\n", i);
 	return (0);
 }
